Eliminate parser dependency on config.
commands.runners depends on config.config to get aliases and the
best_match setting, while config.config depends on commands.runners to
split commands. Eliminate this circular dependency by:

- Moving CommandParser from commands.runners into its a new
  commands.parser module
- Pass aliases/config settings into CommandParser as arguments
