+++ tests/unit/commands/test_parser.py
@@ -17,12 +17,12 @@
 # You should have received a copy of the GNU General Public License
 # along with qutebrowser.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Tests for qutebrowser.commands.runners."""
+"""Tests for qutebrowser.commands.parser."""
 
 import pytest
 
 from qutebrowser.misc import objects
-from qutebrowser.commands import runners, cmdexc
+from qutebrowser.commands import parser, cmdexc
 
 
 class TestCommandParser:
@@ -35,26 +35,26 @@ class TestCommandParser:
         Args:
             cmdline_test: A pytest fixture which provides testcases.
         """
-        parser = runners.CommandParser()
+        p = parser.CommandParser()
         if cmdline_test.valid:
-            parser.parse_all(cmdline_test.cmd, aliases=False)
+            p.parse_all(cmdline_test.cmd)
         else:
             with pytest.raises(cmdexc.NoSuchCommandError):
-                parser.parse_all(cmdline_test.cmd, aliases=False)
+                p.parse_all(cmdline_test.cmd)
 
     def test_parse_all_with_alias(self, cmdline_test, monkeypatch,
                                   config_stub):
         if not cmdline_test.cmd:
             pytest.skip("Empty command")
 
-        config_stub.val.aliases = {'alias_name': cmdline_test.cmd}
+        aliases = {'alias_name': cmdline_test.cmd}
 
-        parser = runners.CommandParser()
+        p = parser.CommandParser()
         if cmdline_test.valid:
-            assert len(parser.parse_all("alias_name")) > 0
+            assert len(p.parse_all("alias_name", aliases=aliases)) > 0
         else:
             with pytest.raises(cmdexc.NoSuchCommandError):
-                parser.parse_all("alias_name")
+                p.parse_all("alias_name", aliases=aliases)
 
     @pytest.mark.parametrize('command', ['', ' '])
     def test_parse_empty_with_alias(self, command):
@@ -63,31 +63,31 @@ class TestCommandParser:
         See https://github.com/qutebrowser/qutebrowser/issues/1690
         and https://github.com/qutebrowser/qutebrowser/issues/1773
         """
-        parser = runners.CommandParser()
+        p = parser.CommandParser()
         with pytest.raises(cmdexc.NoSuchCommandError):
-            parser.parse_all(command)
+            p.parse_all(command, aliases={"foo": "bar"})
 
     @pytest.mark.parametrize('command, name, args', [
         ("set-cmd-text -s :open", "set-cmd-text", ["-s", ":open"]),
         ("set-cmd-text :open {url:pretty}", "set-cmd-text",
-            [":open {url:pretty}"]),
+         [":open {url:pretty}"]),
         ("set-cmd-text -s :open -t", "set-cmd-text", ["-s", ":open -t"]),
         ("set-cmd-text :open -t -r {url:pretty}", "set-cmd-text",
-            [":open -t -r {url:pretty}"]),
+         [":open -t -r {url:pretty}"]),
         ("set-cmd-text -s :open -b", "set-cmd-text", ["-s", ":open -b"]),
         ("set-cmd-text :open -b -r {url:pretty}", "set-cmd-text",
-            [":open -b -r {url:pretty}"]),
+         [":open -b -r {url:pretty}"]),
         ("set-cmd-text -s :open -w", "set-cmd-text",
-            ["-s", ":open -w"]),
+         ["-s", ":open -w"]),
         ("set-cmd-text :open -w {url:pretty}", "set-cmd-text",
-            [":open -w {url:pretty}"]),
+         [":open -w {url:pretty}"]),
         ("set-cmd-text /", "set-cmd-text", ["/"]),
         ("set-cmd-text ?", "set-cmd-text", ["?"]),
         ("set-cmd-text :", "set-cmd-text", [":"]),
     ])
     def test_parse_result(self, command, name, args):
-        parser = runners.CommandParser()
-        result = parser.parse_all(command, aliases=False)[0]
+        p = parser.CommandParser()
+        result = p.parse_all(command)[0]
         assert result.cmd.name == name
         assert result.args == args
 
@@ -110,8 +110,8 @@ class TestCompletions:
 
         The same with it being disabled is tested by test_parse_all.
         """
-        parser = runners.CommandParser(partial_match=True)
-        result = parser.parse('on')
+        p = parser.CommandParser(partial_match=True)
+        result = p.parse('on')
         assert result.cmd.name == 'one'
 
     def test_dont_use_best_match(self, config_stub):
@@ -120,10 +120,10 @@ class TestCompletions:
         Should raise NoSuchCommandError
         """
         config_stub.val.completion.use_best_match = False
-        parser = runners.CommandParser(partial_match=True)
+        p = parser.CommandParser(partial_match=True)
 
         with pytest.raises(cmdexc.NoSuchCommandError):
-            parser.parse('tw')
+            p.parse('tw')
 
     def test_use_best_match(self, config_stub):
         """Test multiple completion options with use_best_match set to true.
@@ -131,7 +131,7 @@ class TestCompletions:
         The resulting command should be the best match
         """
         config_stub.val.completion.use_best_match = True
-        parser = runners.CommandParser(partial_match=True)
+        p = parser.CommandParser(partial_match=True)
 
-        result = parser.parse('tw')
+        result = p.parse('tw', best_match=True)
         assert result.cmd.name == 'two'
